+++ gym_donkeycar/envs/donkey_env.py
@@ -19,14 +19,24 @@ logger = logging.getLogger(__name__)
 
 
 def supply_defaults(conf: Dict[str, Any]) -> None:
+    """
+    Update the config dictonnary
+    with defaults when values are missing.
+
+    :param conf: The user defined config dict,
+        passed to the environment constructor.
+    """
     defaults = [
         ("start_delay", 5.0),
-        ("max_cte", 5.0),
+        ("max_cte", 8.0),
         ("frame_skip", 1),
         ("cam_resolution", (120, 160, 3)),
         ("log_level", logging.INFO),
         ("host", "localhost"),
         ("port", 9091),
+        ("steer_limit", 1.0),
+        ("throttle_min", 0.0),
+        ("throttle_max", 1.0),
     ]
 
     for key, val in defaults:
@@ -46,10 +56,6 @@ class DonkeyEnv(gym.Env):
     metadata = {"render.modes": ["human", "rgb_array"]}
 
     ACTION_NAMES: List[str] = ["steer", "throttle"]
-    STEER_LIMIT_LEFT: float = -1.0
-    STEER_LIMIT_RIGHT: float = 1.0
-    THROTTLE_MIN: float = 0.0
-    THROTTLE_MAX: float = 1.0
     VAL_PER_PIXEL: int = 255
 
     def __init__(self, level: str, conf: Optional[Dict[str, Any]] = None):
@@ -66,7 +72,7 @@ class DonkeyEnv(gym.Env):
         supply_defaults(conf)
 
         # set logging level
-        logging.basicConfig(level=conf["log_level"])  # pytype: disable=key-error
+        logging.basicConfig(level=conf["log_level"])
 
         logger.debug("DEBUG ON")
         logger.debug(conf)
@@ -84,10 +90,12 @@ class DonkeyEnv(gym.Env):
         # start simulation com
         self.viewer = DonkeyUnitySimContoller(conf=conf)
 
+        # Note: for some RL algorithms, it would be better to normalize the action space to [-1, 1]
+        # and then rescale to proper limtis
         # steering and throttle
         self.action_space = spaces.Box(
-            low=np.array([self.STEER_LIMIT_LEFT, self.THROTTLE_MIN]),
-            high=np.array([self.STEER_LIMIT_RIGHT, self.THROTTLE_MAX]),
+            low=np.array([-float(conf["steer_limit"]), float(conf["throttle_min"])]),
+            high=np.array([float(conf["steer_limit"]), float(conf["throttle_max"])]),
             dtype=np.float32,
         )
 
@@ -98,7 +106,7 @@ class DonkeyEnv(gym.Env):
         self.seed()
 
         # Frame Skipping
-        self.frame_skip = conf["frame_skip"]  # pytype: disable=key-error
+        self.frame_skip = conf["frame_skip"]
 
         # wait until the car is loaded in the scene
         self.viewer.wait_until_loaded()
@@ -129,9 +137,13 @@ class DonkeyEnv(gym.Env):
         return observation, reward, done, info
 
     def reset(self) -> np.ndarray:
+        # Activate hand brake, so the car does not move
+        self.viewer.handler.send_control(0, 0, 1.0)
+        time.sleep(0.1)
         self.viewer.reset()
+        self.viewer.handler.send_control(0, 0, 1.0)
+        time.sleep(0.1)
         observation, reward, done, info = self.viewer.observe()
-        time.sleep(1)
         return observation
 
     def render(self, mode: str = "human", close: bool = False) -> Optional[np.ndarray]:
@@ -149,54 +161,54 @@ class DonkeyEnv(gym.Env):
 
 class GeneratedRoadsEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(GeneratedRoadsEnv, self).__init__(level="generated_road", *args, **kwargs)
+        super().__init__(level="generated_road", *args, **kwargs)
 
 
 class WarehouseEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(WarehouseEnv, self).__init__(level="warehouse", *args, **kwargs)
+        super().__init__(level="warehouse", *args, **kwargs)
 
 
 class AvcSparkfunEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(AvcSparkfunEnv, self).__init__(level="sparkfun_avc", *args, **kwargs)
+        super().__init__(level="sparkfun_avc", *args, **kwargs)
 
 
 class GeneratedTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(GeneratedTrackEnv, self).__init__(level="generated_track", *args, **kwargs)
+        super().__init__(level="generated_track", *args, **kwargs)
 
 
 class MountainTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(MountainTrackEnv, self).__init__(level="mountain_track", *args, **kwargs)
+        super().__init__(level="mountain_track", *args, **kwargs)
 
 
 class RoboRacingLeagueTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(RoboRacingLeagueTrackEnv, self).__init__(level="roboracingleague_1", *args, **kwargs)
+        super().__init__(level="roboracingleague_1", *args, **kwargs)
 
 
 class WaveshareEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(WaveshareEnv, self).__init__(level="waveshare", *args, **kwargs)
+        super().__init__(level="waveshare", *args, **kwargs)
 
 
 class MiniMonacoEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(MiniMonacoEnv, self).__init__(level="mini_monaco", *args, **kwargs)
+        super().__init__(level="mini_monaco", *args, **kwargs)
 
 
 class WarrenTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(WarrenTrackEnv, self).__init__(level="warren", *args, **kwargs)
+        super().__init__(level="warren", *args, **kwargs)
 
 
 class ThunderhillTrackEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(ThunderhillTrackEnv, self).__init__(level="thunderhill", *args, **kwargs)
+        super().__init__(level="thunderhill", *args, **kwargs)
 
 
 class CircuitLaunchEnv(DonkeyEnv):
     def __init__(self, *args, **kwargs):
-        super(CircuitLaunchEnv, self).__init__(level="circuit_launch", *args, **kwargs)
+        super().__init__(level="circuit_launch", *args, **kwargs)